# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeResumesRequest {
  /** Base64-encoded job description PDF */
  jobDescriptionBase64: string;
  /** Original filename of the job description */
  jobDescriptionName?: string;
  /** Array of base64-encoded resume PDFs */
  resumesBase64: string[];
  /** Original filenames of the resumes */
  resumeNames?: string[];
}

export interface CandidateResult {
  /** Candidate name */
  name: string;
  /** Candidate type or role category */
  type: string;
  /** Candidate email address */
  email: string;
  /** Candidate contact number */
  contact: string;
  /** Top three skills comma-separated */
  topThreeSkills: string;
  /** Brief summary of the candidate */
  summary: string;
  /** Strengths of the candidate */
  pros: string;
  /** Weaknesses or gaps of the candidate */
  cons: string;
  /** Rating from 0 to 10 */
  rating: number;
  /** Hiring decision Selected Review or Rejected */
  decision: string;
  /** Match score from 0 to 100 */
  matchScore: number;
  /** Skills that match the job description */
  matchingSkills: string[];
  /** Required skills the candidate is missing */
  missingSkills: string[];
}

export interface AnalyzeResumesResponse {
  results: CandidateResult[];
}

export interface ErrorResponse {
  error: string;
}