# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Send base64-encoded PDF files for analysis
 * @summary Analyze resumes against a job description
 */
export const AnalyzeResumesBody = zod.object({
  jobDescriptionBase64: zod
    .string()
    .describe("Base64-encoded job description PDF"),
  jobDescriptionName: zod
    .string()
    .optional()
    .describe("Original filename of the job description"),
  resumesBase64: zod
    .array(zod.string())
    .describe("Array of base64-encoded resume PDFs"),
  resumeNames: zod
    .array(zod.string())
    .optional()
    .describe("Original filenames of the resumes"),
});

export const AnalyzeResumesResponse = zod.object({
  results: zod.array(
    zod.object({
      name: zod.string().describe("Candidate name"),
      type: zod.string().describe("Candidate type or role category"),
      email: zod.string().describe("Candidate email address"),
      contact: zod.string().describe("Candidate contact number"),
      topThreeSkills: zod.string().describe("Top three skills comma-separated"),
      summary: zod.string().describe("Brief summary of the candidate"),
      pros: zod.string().describe("Strengths of the candidate"),
      cons: zod.string().describe("Weaknesses or gaps of the candidate"),
      rating: zod.number().describe("Rating from 0 to 10"),
      decision: zod
        .string()
        .describe("Hiring decision Selected Review or Rejected"),
      matchScore: zod.number().describe("Match score from 0 to 100"),
      matchingSkills: zod
        .array(zod.string())
        .describe("Skills that match the job description"),
      missingSkills: zod
        .array(zod.string())
        .describe("Required skills the candidate is missing"),
    }),
  ),
});