# lib/api-zod/src/generated/types/analyzeResumesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeResumesRequest {
  /** Base64-encoded job description PDF */
  jobDescriptionBase64: string;
  /** Original filename of the job description */
  jobDescriptionName?: string;
  /** Array of base64-encoded resume PDFs */
  resumesBase64: string[];
  /** Original filenames of the resumes */
  resumeNames?: string[];
}