# lib/api-zod/src/generated/types/analyzeResumesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateResult } from "./candidateResult";

export interface AnalyzeResumesResponse {
  results: CandidateResult[];
}